Adding main
This will be the main file for our final project, renamed from layout. I have to look into how to abstract a lot of this since this what a requirement to run pygbag. I opted to use pygbag instead of something like flask since it was made for pygame.

Added:

- asyncio library for use with pygbag. Our main loop contains the asyncio.sleep(0) function (also a requirement). we call our main at the end of the script.
- checkBoardWin function with static variables for bool checks upon recursion.

Refactored:
- Restructured a bit of our main loop for easier use of logic
- Completely overhauled the checkWIn function, a lot of the logic was incorrect. I left multiple conditional/returns in case we want to bring back line draws for anything
- Updated boardDraw for better logic use cases with checkBoardWin and checkWIn functions