Reworked main
Added fill array for board->square relation instead of using rows and columns, may rework this later on to better fit in with dicts

Used fill array to handle conditional for when a board is full but not won. We can also use this for our win condition later on.